rename to_object to to_struct
