+++ lib/representable/object.rb
@@ -3,8 +3,6 @@ require 'representable/object/binding'
 
 module Representable
   module Object
-    autoload :Collection, 'representable/object/collection'
-
     def self.included(base)
       base.class_eval do
         include Representable
@@ -15,23 +13,8 @@ module Representable
 
 
     module ClassMethods
-      def format_engine
-        Representable::Object
-      end
-
       def collection_representer_class
         Collection
-      end
-
-      def cache_struct
-        @represented_struct ||= Struct.new(*representable_attrs.keys.map(&:to_sym))
-      end
-
-      def cache_wrapper_struct(wrap:)
-        struct_name = :"@_wrapper_struct_#{wrap}"
-        return instance_variable_get(struct_name) if instance_variable_defined?(struct_name)
-
-        instance_variable_set(struct_name, Struct.new(wrap))
       end
     end
 
@@ -40,15 +23,7 @@ module Representable
     end
 
     def to_object(options={}, binding_builder=Binding)
-      represented_struct = self.class.cache_struct
-
-      object = create_representation_with(represented_struct.new, options, binding_builder)
-      return object if options[:wrap] == false
-      return object unless (wrap = options[:wrap] || representation_wrap(options))
-
-      wrapper_struct = self.class.cache_wrapper_struct(wrap: wrap.to_sym)
-      wrapper_struct.new(object)
+      create_representation_with(nil, options, binding_builder)
     end
-
   end
 end
