+++ test/examples/object.rb
@@ -5,8 +5,8 @@ require "pp"
 
 source = OpenStruct.new(
   name: "30 Years Live", songs: [
-  OpenStruct.new(id: 1, title: "Dear Beloved"), OpenStruct.new(id: 2, title: "Fuck Armageddon")
-]
+    OpenStruct.new(id: 1, title: "Dear Beloved"), OpenStruct.new(id: 2, title: "Fuck Armageddon")
+  ]
 )
 
 require "representable/object"
@@ -26,30 +26,3 @@ Song = Struct.new(:title)
 target = Album.new
 
 AlbumRepresenter.new(target).from_object(source)
-
-# Representing to_object:
-
-class Animal
-  attr_accessor :name, :age, :species
-  def initialize(name, age, species)
-    @name = name
-    @age = age
-    @species = species
-  end
-end
-
-require "representable/object"
-
-class AnimalRepresenter < Representable::Decorator
-  include Representable::Object
-
-  property :name, getter: ->(represented:, **) { represented.name.upcase }
-  property :age
-end
-
-animal = Animal.new('Smokey',12,'c')
-animal_repr = AnimalRepresenter.new(animal).to_object(wrap: "wrapper")
-
-animal_array = [Animal.new('Shepard',22,'s'),Animal.new('Pickle',12,'c'),Animal.new('Rodgers',55,'e')]
-array_repr = AnimalRepresenter.for_collection.new(animal_array).to_object
-
